requirements.txt
Django==4.1.2
# mysqlclient==2.1.1
# mysql-connector-python==8.0.25
psycopg2==2.9.3
django-cors-headers==3.13.0
six==1.16.0  # via django-bootstrap-breadcrumbs
django-crispy-forms==1.14.0
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
WTForms-Dynamic-Fields==0.1a3
Flask-Session==0.3.2
bootstrap4==0.1.0
kwargs==1.0.1
docutils==0.19
whitenoise==6.2.0
Pillow==9.2
zeep==4.1.0
django-hitcount==1.3.5
django-staticfiles==1.2.1
use-breadcrumbs==0.2
djoser==2.0.5
djangorestframework-oauth==1.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1